+++ tests/executor/preprocessor/test_integration.py
@@ -1,15 +1,30 @@
 import os
 
 import hubble
+import pytest
 from docarray import Document, DocumentArray
-from jina import Flow
+from jina import Client
 
-from now.executor.gateway.now_gateway import NOWGateway
-from now.executor.preprocessor import NOWPreprocessor
+from now.common.options import construct_app
+from now.constants import Apps
 from now.now_dataclasses import UserInput
 
 
-def test_search_app(resources_folder_path, tmpdir, mm_dataclass):
+@pytest.fixture
+def preprocessor_test(resources_folder_path, mm_dataclass):
+    user_input = UserInput()
+    user_input.user_emails = ['team-now@jina.ai']
+    user_input.admin_emails = []
+    user_input.api_key = []
+    user_input.jwt = {'token': hubble.get_token()}
+    user_input.app_instance = construct_app(Apps.SEARCH_APP)
+
+    def func(*args, **kwargs):
+        preprocess_config = user_input.app_instance.preprocessor_stub(True)
+        preprocess_config.pop('needs')
+        return [preprocess_config]
+
+    user_input.app_instance.get_executor_stubs = func
     docs = DocumentArray(
         [
             Document(mm_dataclass(text_field='test')),
@@ -22,36 +37,29 @@ def test_search_app(resources_folder_path, tmpdir, mm_dataclass):
             ),
         ]
     )
-    # this is needed as using http compression leads to a tensor of dtype uint8 is received as int64
     docs[1].video_field.load_uri_to_blob(timeout=10)
-    user_input = UserInput()
-    user_input.user_emails = ['team-now@jina.ai']
-    user_input.admin_emails = []
-    user_input.api_key = []
-    user_input.jwt = {'token': hubble.get_token()}
+    return docs, user_input
 
-    with Flow().config_gateway(
-        uses=NOWGateway,
-        protocol=['http', 'grpc'],
-        port=[8081, 8085],
-        env={'JINA_LOG_LEVEL': 'DEBUG'},
-        uses_with={
-            'user_input_dict': user_input.__dict__,
-            'm2m_token': os.environ['M2M_TOKEN'],
-        },
-    ).add(uses=NOWPreprocessor, workspace=tmpdir, env={'JINA_LOG_LEVEL': 'DEBUG'}) as f:
-        for endpoint in ['index', 'search']:
-            result = f.post(
-                on=f'/{endpoint}',
-                inputs=docs,
-                show_progress=True,
-                headers={
-                    'authorization': f'token {hubble.get_token()}',
-                    'Content-Type': 'application/json',
-                },
-            )
 
-            assert len(result) == 2
-            assert result[0].text == ''
-            assert result[0].chunks[0].chunks[0].text == 'test'
-            assert result[1].chunks[0].chunks[0].blob
+@pytest.mark.parametrize(
+    'get_flow',
+    ['preprocessor_test'],
+    indirect=True,
+)
+def test_search_app(get_flow):
+    docs, user_input = get_flow
+    client = Client(host='grpc://localhost:8085')
+
+    # this is needed as using http compression leads to a tensor of dtype uint8 is received as int64
+    for endpoint in ['index', 'search']:
+        result = client.post(
+            on=f'/{endpoint}',
+            inputs=docs,
+            show_progress=True,
+            metadata=(('authorization', hubble.get_token()),),
+        )
+
+        assert len(result) == 2
+        assert result[0].text == ''
+        assert result[0].chunks[0].chunks[0].text == 'test'
+        assert result[1].chunks[0].chunks[0].blob
